# Conflicts:
#	.dockerignore
#	.env.example
#	.gitignore
#	Dockerfile.web
#	Makefile
#	apps/authentication/api_views.py
#	apps/authentication/serializers.py
#	apps/authentication/urls.py
#	apps/chat/migrations/0001_initial.py
#	apps/content/migrations/0001_initial.py
#	apps/teams/migrations/0003_team_billing_details_last_changed_team_customer_and_more.py
#	apps/users/migrations/0002_customuser_language_customuser_timezone_and_more.py
#	apps/web/migrations/0001_initial.py
#	bh_crypto/settings.py
#	bh_crypto/settings_production.py
#	bh_reggie/urls.py
#	deploy/1_env.google.example
#	docker_startup.sh
#	frontend/src/api/utils.tsx
#	frontend/src/auth/authprovider.tsx
#	package.json
#	pegasus-config.yaml
#	requirements/dev-requirements.txt
#	requirements/prod-requirements.txt
#	requirements/requirements.in
#	requirements/requirements.txt
#	static/css/site-base.css
#	static/css/site-tailwind.css
#	static/js/app-bundle.js
#	static/js/chat-bundle.js
#	static/js/dashboard-bundle.js
#	static/js/edit-team-bundle.js
#	static/js/site-base-bundle.js
#	static/js/site-bundle.js
#	static/js/site-tailwind-bundle.js
#	static/js/teams-bundle.js

+++ apps/users/signals.py
@@ -32,7 +32,7 @@ def update_user_email(sender, request, email_address, **kwargs):
 def _notify_admins_of_signup(user):
     mail_admins(
         f"Yowsers, someone signed up for {settings.PROJECT_METADATA['NAME']}!",
-        "Email: {}".format(user.email),
+        f"Email: {user.email}",
         fail_silently=True,
     )
 
@@ -47,13 +47,11 @@ def remove_old_profile_picture_on_change(sender, instance, **kwargs):
     except sender.DoesNotExist:
         return False
 
-    if old_file and old_file.name != instance.avatar.name:
-        if default_storage.exists(old_file.name):
-            default_storage.delete(old_file.name)
+    if old_file and old_file.name != instance.avatar.name and default_storage.exists(old_file.name):
+        default_storage.delete(old_file.name)
 
 
 @receiver(post_delete, sender=CustomUser)
 def remove_profile_picture_on_delete(sender, instance, **kwargs):
-    if instance.avatar:
-        if default_storage.exists(instance.avatar.name):
-            default_storage.delete(instance.avatar.name)
+    if instance.avatar and default_storage.exists(instance.avatar.name):
+        default_storage.delete(instance.avatar.name)
